requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
click==8.3.0
colorama==0.4.6
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.119.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
passlib==1.7.4
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.3
pydantic_core==2.41.4
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
yarl==1.22.0
